+++ src/Infrastructure/DTO/DomainData.php
@@ -4,34 +4,20 @@ declare(strict_types=1);
 
 namespace hiqdev\rdap\core\Infrastructure\DTO;
 
+use DateTimeImmutable;
+use DateTimeZone;
+
 final class DomainData
 {
-    /** @var string */
-    private $handle;
-
-    /** @var string|null comma-separated EPP status names */
-    private $statuses;
-
-    /** @var string|null comma-separated nameserver host names */
-    private $nameservers;
-
-    /** @var string */
-    private $creationDate;
-
-    /** @var string */
-    private $updatedDate;
-
-    /** @var string registrar expiration date */
-    private $registrarExpiration;
-
-    /** @var string registry expiration date */
-    private $expiration;
-
-    /** @var bool */
-    private $whoisProtected;
-
-    /** @var bool */
-    private $delegationSigned;
+    private string $handle;
+    private ?string $statuses;
+    private ?string $nameservers;
+    private DateTimeImmutable $creationDate;
+    private DateTimeImmutable $updatedDate;
+    private DateTimeImmutable $registrarExpiration;
+    private DateTimeImmutable $expiration;
+    private bool $whoisProtected;
+    private bool $delegationSigned;
 
     public function __construct(
         string $handle,
@@ -44,13 +30,14 @@ final class DomainData
         bool $whoisProtected,
         bool $delegationSigned
     ) {
+        $utc = new DateTimeZone('UTC');
         $this->handle              = $handle;
         $this->statuses            = $statuses;
         $this->nameservers         = $nameservers;
-        $this->creationDate        = $creationDate;
-        $this->updatedDate         = $updatedDate;
-        $this->registrarExpiration = $registrarExpiration;
-        $this->expiration          = $expiration;
+        $this->creationDate        = new DateTimeImmutable($creationDate, $utc);
+        $this->updatedDate         = new DateTimeImmutable($updatedDate, $utc);
+        $this->registrarExpiration = new DateTimeImmutable($registrarExpiration, $utc);
+        $this->expiration          = new DateTimeImmutable($expiration, $utc);
         $this->whoisProtected      = $whoisProtected;
         $this->delegationSigned    = $delegationSigned;
     }
@@ -58,10 +45,10 @@ final class DomainData
     public function getHandle(): string { return $this->handle; }
     public function getStatuses(): ?string { return $this->statuses; }
     public function getNameservers(): ?string { return $this->nameservers; }
-    public function getCreationDate(): string { return $this->creationDate; }
-    public function getUpdatedDate(): string { return $this->updatedDate; }
-    public function getRegistrarExpiration(): string { return $this->registrarExpiration; }
-    public function getExpiration(): string { return $this->expiration; }
+    public function getCreationDate(): DateTimeImmutable { return $this->creationDate; }
+    public function getUpdatedDate(): DateTimeImmutable { return $this->updatedDate; }
+    public function getRegistrarExpiration(): DateTimeImmutable { return $this->registrarExpiration; }
+    public function getExpiration(): DateTimeImmutable { return $this->expiration; }
     public function isWhoisProtected(): bool { return $this->whoisProtected; }
     public function isDelegationSigned(): bool { return $this->delegationSigned; }
 }
