+++ src/Infrastructure/Provider/NoticeBuilder.php
@@ -1,0 +1,32 @@
+<?php
+
+declare(strict_types=1);
+
+namespace hiqdev\rdap\core\Infrastructure\Provider;
+
+use hiqdev\rdap\core\Domain\ValueObject\Link;
+use hiqdev\rdap\core\Domain\ValueObject\Notice;
+
+final class NoticeBuilder
+{
+    public function __construct(private array $noticesConfig)
+    {
+    }
+
+    /**
+     * @return Notice[]
+     */
+    public function build(string $currentUrl): array
+    {
+        $notices = [];
+        foreach ($this->noticesConfig as $config) {
+            $link = new Link($config['link']['href']);
+            $link->setType('text/html');
+            $link->setValue($currentUrl);
+            $link->setRel($config['link']['rel']);
+            $notices[] = new Notice($config['title'], [$config['description']], [$link]);
+        }
+
+        return $notices;
+    }
+}
